Update golangci-lint action. (#42)
* Update golangci-lint action.

* Update comproc for Go 1.18.
