+++ .github/workflows/go.yml
@@ -33,7 +33,10 @@ jobs:
     runs-on: windows-latest
     steps:
     - uses: actions/checkout@v2
-    - name: Lint
-      uses: golangci/golangci-lint-action@v2
+    - uses: actions/setup-go@v3
       with:
-        version: v1.44
+        go-version: 1.18
+    - name: Lint
+      uses: golangci/golangci-lint-action@v3
+      with:
+        version: v1.45
