+++ pkg/edge/comproc.go
@@ -1,0 +1,13 @@
+package edge
+
+import (
+	"golang.org/x/sys/windows"
+)
+
+// ComProc stores a COM procedure.
+type ComProc uintptr
+
+// NewComProc creates a new COM proc from a Go function.
+func NewComProc(fn interface{}) ComProc {
+	return ComProc(windows.NewCallback(fn))
+}
